analytics/pyproject.toml
[project]
name = "fa-qi-analytics"
version = "0.1.0"
description = "Ingestion and engines for Financial Advisor quantitative intelligence"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "sqlalchemy>=2.0,<3",
  "psycopg[binary]>=3.1,<4",
  "httpx>=0.27,<1",
  "python-dotenv>=1.0,<2",
  "numpy>=1.26,<3",
]

[project.optional-dependencies]
dev = ["pytest>=8", "ruff>=0.6"]

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["qi*"]

[tool.ruff]
line-length = 100
target-version = "py311"
